openWebsite function in a separate file.
