+++ push.go
@@ -6,9 +6,7 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-	"os/exec"
 	"regexp"
-	"runtime"
 	"time"
 
 	"gopkg.in/alecthomas/kingpin.v2"
@@ -375,20 +373,4 @@ func getSessionID(appPath string) (string, error) {
 	}
 
 	return settings.SessionID, nil
-}
-
-func openWebsite(url string) error {
-	var err error
-	switch runtime.GOOS {
-	case "linux":
-		err = exec.Command("xdg-open", url).Start()
-	case "windows":
-		err = exec.Command("rundll32", "url.dll,FileProtocolHandler", url).Start()
-	case "darwin":
-		err = exec.Command("open", url).Start()
-	default:
-		err = fmt.Errorf("unsupported platform")
-	}
-
-	return err
 }
